[#492] Address PR feedback
