+++ src/electron/shared/dto/DailySurveyDto.ts
@@ -1,5 +1,13 @@
 import type { DailySurveySamplingType, ExperienceSamplingAnswerType } from '../StudyConfiguration';
 
+export interface DailySurveyResponseInput {
+  question: string;
+  answerType: ExperienceSamplingAnswerType;
+  responseOptions: string | null;
+  response: string | null;
+  skipped: boolean;
+}
+
 export default interface DailySurveyDto {
   id: string;
   promptedAt: Date;
@@ -7,7 +15,6 @@ export default interface DailySurveyDto {
   question: string;
   answerType: ExperienceSamplingAnswerType;
   responseOptions: string | null;
-  scale: number | null;
   response: string | null;
   skipped: boolean;
   createdAt: Date;
